+++ library/src/device/kernels/callback.h
@@ -37,41 +37,6 @@ struct UserCallbacks
     size_t store_cb_lds_bytes = 0;
 };
 
-// callback function types
-template <typename T>
-struct callback_type;
-
-template <>
-struct callback_type<float>
-{
-    typedef float (*load)(float* data, size_t offset, void* cbdata, void* sharedMem);
-    typedef void (*store)(float* data, size_t offset, float element, void* cbdata, void* sharedMem);
-};
-
-template <>
-struct callback_type<float2>
-{
-    typedef float2 (*load)(float2* data, size_t offset, void* cbdata, void* sharedMem);
-    typedef void (*store)(
-        float2* data, size_t offset, float2 element, void* cbdata, void* sharedMem);
-};
-
-template <>
-struct callback_type<double>
-{
-    typedef double (*load)(double* data, size_t offset, void* cbdata, void* sharedMem);
-    typedef void (*store)(
-        double* data, size_t offset, double element, void* cbdata, void* sharedMem);
-};
-
-template <>
-struct callback_type<double2>
-{
-    typedef double2 (*load)(double2* data, size_t offset, void* cbdata, void* sharedMem);
-    typedef void (*store)(
-        double2* data, size_t offset, double2 element, void* cbdata, void* sharedMem);
-};
-
 // default callback implementations that just do simple load/store
 template <typename T>
 __device__ T load_cb_default(T* data, size_t offset, void* cbdata, void* sharedMem)
@@ -84,6 +49,53 @@ __device__ void store_cb_default(T* data, size_t offset, T element, void* cbdata
 {
     data[offset] = element;
 }
+
+// callback function types
+template <typename T>
+struct callback_type;
+
+template <>
+struct callback_type<float>
+{
+    typedef float (*load)(float* data, size_t offset, void* cbdata, void* sharedMem);
+    typedef void (*store)(float* data, size_t offset, float element, void* cbdata, void* sharedMem);
+};
+
+__device__ auto load_cb_default_float  = load_cb_default<float>;
+__device__ auto store_cb_default_float = store_cb_default<float>;
+
+template <>
+struct callback_type<float2>
+{
+    typedef float2 (*load)(float2* data, size_t offset, void* cbdata, void* sharedMem);
+    typedef void (*store)(
+        float2* data, size_t offset, float2 element, void* cbdata, void* sharedMem);
+};
+
+__device__ auto load_cb_default_float2  = load_cb_default<float2>;
+__device__ auto store_cb_default_float2 = store_cb_default<float2>;
+
+template <>
+struct callback_type<double>
+{
+    typedef double (*load)(double* data, size_t offset, void* cbdata, void* sharedMem);
+    typedef void (*store)(
+        double* data, size_t offset, double element, void* cbdata, void* sharedMem);
+};
+
+__device__ auto load_cb_default_double  = load_cb_default<double>;
+__device__ auto store_cb_default_double = store_cb_default<double>;
+
+template <>
+struct callback_type<double2>
+{
+    typedef double2 (*load)(double2* data, size_t offset, void* cbdata, void* sharedMem);
+    typedef void (*store)(
+        double2* data, size_t offset, double2 element, void* cbdata, void* sharedMem);
+};
+
+__device__ auto load_cb_default_double2  = load_cb_default<double2>;
+__device__ auto store_cb_default_double2 = store_cb_default<double2>;
 
 enum struct CallbackType
 {
@@ -98,7 +110,7 @@ template <typename T, CallbackType cbtype>
 static __device__ typename callback_type<T>::load get_load_cb(void* ptr)
 {
 #ifdef ROCFFT_CALLBACKS_ENABLED
-    if(cbtype == CallbackType::USER_LOAD_STORE && ptr != nullptr)
+    if(cbtype == CallbackType::USER_LOAD_STORE)
         return reinterpret_cast<typename callback_type<T>::load>(ptr);
 #endif
     return load_cb_default<T>;
@@ -108,7 +120,7 @@ template <typename T, CallbackType cbtype>
 static __device__ typename callback_type<T>::store get_store_cb(void* ptr)
 {
 #ifdef ROCFFT_CALLBACKS_ENABLED
-    if(cbtype == CallbackType::USER_LOAD_STORE && ptr != nullptr)
+    if(cbtype == CallbackType::USER_LOAD_STORE)
         return reinterpret_cast<typename callback_type<T>::store>(ptr);
 #endif
     return store_cb_default<T>;